Implement data caching, and do database work on background thread.
With this patch, every event updates a local data cache as well
as commits to the database.  Future patches will use this cache
to read frequently accessed data, only hitting the database
on writes, or occasionally on reads when the system thinks something
may have changed.

Additionally, we now do all database queries on a single background
thread.  This was necesary since previously we would hit the
database from both the UI thread or the proxy thread, and we need
to enforce that all database operations are synchronized.

+++ ffrk-winproxy/GameData/DataBattle.cs
@@ -22,7 +22,7 @@ namespace ffrk_winproxy.GameData
         public bool HasBoss;
 
         [JsonProperty("stamina")]
-        public uint Stamina;
+        public ushort Stamina;
 
         [JsonProperty("is_unlocked")]
         public bool IsUnlocked;
